refactor: websocket store 관련 파일 구조 변경, 파일명 kebab-case에 맞춰 수정 #46, #50
