+++ src/page/DeviceMonitoringPage.vue
@@ -98,7 +98,7 @@
 
 <script setup>
 import { ref, onMounted, watch } from "vue";
-import { useWebSocketStore } from '@/store/websocketStore';
+import { useWebSocketStore } from '@/store';
 import HeaderTemplate from "@/components/HeaderTemplate.vue";
 import DataTable from "primevue/datatable";
 import Column from "primevue/column";
